refactor: move AuthServiceImpl to core/data and rename
This commit moves the `AuthServiceImpl` class from `core/firebase-services` to `core/data` and renames it to `AuthRepositoryImpl`. It also updates the package declaration and implements the `com.casecode.pos.core.domain.repository.AuthRepository` interface.

This change is made to improve the structure of the project by moving the authentication logic to the data layer. By implementing the `AuthRepository` interface, it also ensures consistency and testability.