feat(designsystem): add default value for isSelectableDate in DatePicker
Added a default value to the `isSelectableDate` parameter in the `DatePicker` composable.

This change makes `isSelectableDate` optional and defaults to allowing all dates to be selected if not provided. This improves the usability of the component by removing the need for users to explicitly specify a selection logic if they want to allow all dates.