feat(core-data): add Instant toFormattedDateTimeString extension function
Added an extension function `toFormattedDateTimeString` to `Instant` for formatting it into a user-readable date and time string.

This function formats the `Instant` to a `LocalDateTime` using the current system's default time zone and then formats it to the following format:
"dd MMM yyyy hh:mm a", for example "01 Jan 2024 10:00 AM".